django_portal/requirements.txt
# Django Core
Django>=5.0,<6.0
django-environ>=0.11.2
psycopg2-binary>=2.9.0
redis>=5.0.0

# Authentication & Users
django-allauth>=0.57.0
django-crispy-forms>=2.1
crispy-tailwind>=0.9.0

# File Handling & Storage
Pillow>=10.0.0
django-storages>=1.14.0

# PDF Generation
WeasyPrint>=60.0

# Background Tasks
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
flower>=2.0.0

# Forms & Widgets
django-widget-tweaks>=1.4.12

# Development & Debugging
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0

# API & Integration
requests>=2.31.0
xmltodict>=0.13.0

# Security
django-cors-headers>=4.3.0

# Utilities
python-decouple>=3.8
gunicorn>=21.0.0
whitenoise>=6.6.0

# Email
django-anymail>=10.2

# Date/Time
python-dateutil>=2.8.0
pytz>=2023.3

# Validation
django-phonenumber-field>=7.3.0
phonenumbers>=8.13.0

# File processing
python-magic>=0.4.27

# Logging
sentry-sdk>=1.38.0

# Testing (optional, for development)
pytest-django>=4.5.0
pytest-cov>=4.1.0
